[UIMA-6393]: Circular imports break resource manager cache
- Pulled import resolving code out into a separate class
- Pulled import resolving tests out into a separate test class
- Changed log layout
- Use generic import resolver in TypeSystemDescription_impl, FsIndexCollection_impl and TypePriorities_impl
- Deprecate import resolving method allowing the passing "already imported" URLs in all three descriptor types
- Added back setting the source URL on imports if it was null
- Added back using the "already imported" argument of the deprecated resolve method in the generic import resolver - not because it is a good idea, just for backwards compatibility
