+++ uimaj-core/src/main/java/org/apache/uima/resource/metadata/FsIndexCollection.java
@@ -52,7 +52,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * 
    * @return the name of this FS Index Collection, null if none has been specified.
    */
-  public String getName();
+  String getName();
 
   /**
    * Sets the name of this FS Index Collection.
@@ -63,7 +63,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setName(String aName);
+  void setName(String aName);
 
   /**
    * Gets the version number of this FS Index Collection.
@@ -71,7 +71,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @return the version number of this FS Index Collection, as a String, null if none has been
    *         specified.
    */
-  public String getVersion();
+  String getVersion();
 
   /**
    * Sets the version number of this FS Index Collection.
@@ -82,14 +82,14 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setVersion(String aVersion);
+  void setVersion(String aVersion);
 
   /**
    * Gets the description of this FS Index Collection.
    * 
    * @return the description of this FS Index Collection, null if none has been specified.
    */
-  public String getDescription();
+  String getDescription();
 
   /**
    * Sets the description of this FS Index Collection.
@@ -100,14 +100,14 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setDescription(String aDescription);
+  void setDescription(String aDescription);
 
   /**
    * Gets the vendor of this FS Index Collection.
    * 
    * @return the vendor of this FS Index Collection, as a String
    */
-  public String getVendor();
+  String getVendor();
 
   /**
    * Sets the vendor of this FS Index Collection.
@@ -118,14 +118,14 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setVendor(String aVendor);
+  void setVendor(String aVendor);
 
   /**
    * Gets the imports declared by this FS Index Collection.
    * 
    * @return an array of imports declared by this FS Index Collection.
    */
-  public Import[] getImports();
+  Import[] getImports();
 
   /**
    * Sets the imports declared by this FS Index Collection.
@@ -136,7 +136,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setImports(Import[] aImports);
+  void setImports(Import[] aImports);
 
   /**
    * Retrieves the FS Index descriptions that are part of this collection. These define the indexes
@@ -144,7 +144,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * 
    * @return a description of the FS Indexes that comprise this FS Index Collection
    */
-  public FsIndexDescription[] getFsIndexes();
+  FsIndexDescription[] getFsIndexes();
 
   /**
    * Retrieves the FS Index descriptions that are part of this collection. These define the indexes
@@ -156,7 +156,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void setFsIndexes(FsIndexDescription[] aFSIndexes);
+  void setFsIndexes(FsIndexDescription[] aFSIndexes);
 
   /**
    * Adds an FS Index description to this collection.
@@ -167,7 +167,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void addFsIndex(FsIndexDescription aFsIndexDescription);
+  void addFsIndex(FsIndexDescription aFsIndexDescription);
 
   /**
    * Removes an FS Index description from this collection.
@@ -179,7 +179,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws UIMA_UnsupportedOperationException
    *           if this object is not modifiable
    */
-  public void removeFsIndex(FsIndexDescription aFsIndexDescription);
+  void removeFsIndex(FsIndexDescription aFsIndexDescription);
 
   /**
    * Resolves any import declarations in this FS Index Collection, adding the imported
@@ -190,7 +190,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws InvalidXMLException
    *           if either the import target does not exist or is invalid
    */
-  public void resolveImports() throws InvalidXMLException;
+  void resolveImports() throws InvalidXMLException;
 
   /**
    * Resolves any import declarations in this FS Index Collection, adding the imported
@@ -206,7 +206,7 @@ public interface FsIndexCollection extends MetaDataObject {
    * @throws InvalidXMLException
    *           if either the import target does not exist or is invalid
    */
-  public void resolveImports(ResourceManager aResourceManager) throws InvalidXMLException;
+  void resolveImports(ResourceManager aResourceManager) throws InvalidXMLException;
 
   /**
    * Resolves any import declarations in this FS Index Collection, adding the imported
@@ -225,7 +225,10 @@ public interface FsIndexCollection extends MetaDataObject {
    * 
    * @throws InvalidXMLException
    *           if either the import target does not exist or is invalid
+   * @deprecated Method is deprecated because it should only be used internally. Scheduled for
+   *             removal in UIMA 4.0.
    */
-  public void resolveImports(Collection<String> aAlreadyImportedFsIndexCollectionURLs,
+  @Deprecated
+  void resolveImports(Collection<String> aAlreadyImportedFsIndexCollectionURLs,
           ResourceManager aResourceManager) throws InvalidXMLException;
 }
