+++ uimaj-core/src/main/java/org/apache/uima/resource/metadata/impl/FsIndexCollection_impl.java
@@ -19,32 +19,16 @@
 
 package org.apache.uima.resource.metadata.impl;
 
-import java.io.IOException;
-import java.net.URL;
-import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeSet;
 
-import org.apache.uima.UIMAFramework;
 import org.apache.uima.UIMA_IllegalArgumentException;
 import org.apache.uima.resource.ResourceManager;
-import org.apache.uima.resource.impl.ResourceManager_impl;
 import org.apache.uima.resource.metadata.FsIndexCollection;
 import org.apache.uima.resource.metadata.FsIndexDescription;
 import org.apache.uima.resource.metadata.Import;
 import org.apache.uima.resource.metadata.ResourceMetaData;
 import org.apache.uima.util.InvalidXMLException;
-import org.apache.uima.util.XMLInputSource;
-import org.apache.uima.util.XMLizable;
 
-/**
- * 
- * 
- */
 public class FsIndexCollection_impl extends MetaDataObject_impl implements FsIndexCollection {
 
   private static final long serialVersionUID = -7687383527183197102L;
@@ -66,6 +50,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#getName()
    */
+  @Override
   public String getName() {
     return mName;
   }
@@ -73,6 +58,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#setName(String)
    */
+  @Override
   public void setName(String aName) {
     mName = aName;
   }
@@ -80,6 +66,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#getVersion()
    */
+  @Override
   public String getVersion() {
     return mVersion;
   }
@@ -87,6 +74,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#setVersion(String)
    */
+  @Override
   public void setVersion(String aVersion) {
     mVersion = aVersion;
   }
@@ -94,6 +82,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#getDescription()
    */
+  @Override
   public String getDescription() {
     return mDescription;
   }
@@ -101,6 +90,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#setDescription(String)
    */
+  @Override
   public void setDescription(String aDescription) {
     mDescription = aDescription;
   }
@@ -108,6 +98,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#getVendor()
    */
+  @Override
   public String getVendor() {
     return mVendor;
   }
@@ -115,6 +106,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
   /**
    * @see ResourceMetaData#setVendor(String)
    */
+  @Override
   public void setVendor(String aVendor) {
     mVendor = aVendor;
   }
@@ -124,6 +116,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
    * 
    * @see org.apache.uima.resource.metadata.TypeSystemDescription#getImports()
    */
+  @Override
   public Import[] getImports() {
     // don't allow this to return null
     return (mImports == null) ? Import.EMPTY_IMPORTS : mImports;
@@ -134,6 +127,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
    * 
    * @see org.apache.uima.resource.metadata.TypeSystemDescription#setImports(org.apache.uima.resource.metadata.Import[])
    */
+  @Override
   public void setImports(Import[] aImports) {
     if (aImports == null) {
       throw new UIMA_IllegalArgumentException(UIMA_IllegalArgumentException.ILLEGAL_ARGUMENT,
@@ -147,10 +141,12 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
    * 
    * @see org.apache.uima.resource.metadata.FsIndexCollection#getFsIndexes()
    */
+  @Override
   public FsIndexDescription[] getFsIndexes() {
     // don't allow this to return null
-    if (mFsIndexes == null)
+    if (mFsIndexes == null) {
       mFsIndexes = EMPTY_FS_INDEX_DESCRIPTION_ARRAY;
+    }
     return mFsIndexes;
   }
 
@@ -159,6 +155,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
    * 
    * @see org.apache.uima.resource.metadata.FsIndexCollection#setFsIndexes(org.apache.uima.resource.metadata.FsIndexDescription[])
    */
+  @Override
   public void setFsIndexes(FsIndexDescription[] aFSIndexes) {
     if (aFSIndexes == null) {
       throw new UIMA_IllegalArgumentException(UIMA_IllegalArgumentException.ILLEGAL_ARGUMENT,
@@ -167,6 +164,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
     mFsIndexes = aFSIndexes;
   }
 
+  @Override
   public void addFsIndex(FsIndexDescription aFsIndexDescription) {
     FsIndexDescription[] current = getFsIndexes();
     FsIndexDescription[] newArr = new FsIndexDescription[current.length + 1];
@@ -175,6 +173,7 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
     setFsIndexes(newArr);
   }
 
+  @Override
   public void removeFsIndex(FsIndexDescription aFsIndexDescription) {
     FsIndexDescription[] current = getFsIndexes();
     for (int i = 0; i < current.length; i++) {
@@ -194,96 +193,26 @@ public class FsIndexCollection_impl extends MetaDataObject_impl implements FsInd
    * @see org.apache.uima.resource.metadata.TypeSystemDescription#resolveImports()
    */
   // support multi-threading, avoid object creation if no imports
+  @Override
   public synchronized void resolveImports() throws InvalidXMLException {
-    if (getImports().length == 0) {
-      resolveImports(null, null);
-    } else {
-      resolveImports(new TreeSet<>(), UIMAFramework.newDefaultResourceManager());
-    }
+    resolveImports(null, null);
   }
 
+  @Override
   public synchronized void resolveImports(ResourceManager aResourceManager) throws InvalidXMLException {
-    resolveImports((getImports().length == 0) ? null : new TreeSet<>(), aResourceManager);
+    resolveImports(null, aResourceManager);
   }
 
+  @Deprecated
+  @Override
   public synchronized void resolveImports(Collection<String> aAlreadyImportedFsIndexURLs,
           ResourceManager aResourceManager) throws InvalidXMLException {
-    List<FsIndexDescription> importedIndexes = null;
-    if (getImports().length != 0) {
-      // add our own URL, if known, to the collection of already imported URLs
-      if (getSourceUrl() != null) {
-        aAlreadyImportedFsIndexURLs.add(getSourceUrl().toString());
-      }
-      
-      importedIndexes = new ArrayList<>();
-      Import[] imports = getImports();
-      for (int i = 0; i < imports.length; i++) {
-        // make sure Import's relative path base is set, to allow for users who create
-        // new import objects
-        if (imports[i] instanceof Import_impl) {
-          ((Import_impl) imports[i]).setSourceUrlIfNull(this.getSourceUrl());
-        }
-  
-        URL url = imports[i].findAbsoluteUrl(aResourceManager);
-        if (!aAlreadyImportedFsIndexURLs.contains(url.toString())) {
-          aAlreadyImportedFsIndexURLs.add(url.toString());
-          try {
-            resolveImport(url, aAlreadyImportedFsIndexURLs, importedIndexes, aResourceManager);
-          } catch (IOException e) {
-            throw new InvalidXMLException(InvalidXMLException.IMPORT_FAILED_COULD_NOT_READ_FROM_URL,
-                    new Object[] { url, imports[i].getSourceUrlString() }, e);
-          }
-        }
-      }
-    }
-    // update this object
-    FsIndexDescription[] existingIndexes = this.getFsIndexes();
-    if (existingIndexes == null) {
-      this.setFsIndexes(existingIndexes = FsIndexDescription.EMPTY_FS_INDEX_DESCRIPTIONS);
-    }
-    if (null != importedIndexes) {
-      FsIndexDescription[] newIndexes = new FsIndexDescription[existingIndexes.length
-              + importedIndexes.size()];
-      System.arraycopy(existingIndexes, 0, newIndexes, 0, existingIndexes.length);
-      for (int i = 0; i < importedIndexes.size(); i++) {
-        newIndexes[existingIndexes.length + i] = importedIndexes.get(i);
-      }
-      this.setFsIndexes(newIndexes);
-    }
-    // clear imports
-    this.setImports(Import.EMPTY_IMPORTS);
+    ImportResolver<FsIndexCollection, FsIndexDescription> resolver = new ImportResolver<>(
+            FsIndexCollectionImportResolverAdapter::new);
+    resolver.resolveImports(this, aAlreadyImportedFsIndexURLs, aResourceManager);
   }
 
-  private void resolveImport(URL aURL, Collection<String> aAlreadyImportedFsIndexCollectionURLs,
-          Collection<FsIndexDescription> aResults, ResourceManager aResourceManager) throws InvalidXMLException,
-          IOException {
-    //check the import cache
-    FsIndexCollection desc;    
-    String urlString = aURL.toString();
-    Map<String, XMLizable> importCache = ((ResourceManager_impl)aResourceManager).getImportCache();
-    Map<String, Set<String>> importUrlsCache = ((ResourceManager_impl)aResourceManager).getImportUrlsCache();
-    synchronized(importCache) {
-      XMLizable cachedObject = importCache.get(urlString);
-      if (cachedObject instanceof FsIndexCollection) {
-        desc = (FsIndexCollection)cachedObject;
-        // Add the URLs parsed for this cached object to the list already-parsed (UIMA-5058)
-        aAlreadyImportedFsIndexCollectionURLs.addAll(importUrlsCache.get(urlString));
-      } else {   
-        XMLInputSource input;
-        input = new XMLInputSource(aURL);
-        desc = UIMAFramework.getXMLParser().parseFsIndexCollection(input);
-        TreeSet<String> previouslyImported = new TreeSet<>(aAlreadyImportedFsIndexCollectionURLs);
-        desc.resolveImports(aAlreadyImportedFsIndexCollectionURLs, aResourceManager);
-        importCache.put(urlString, desc);
-        // Save the URLS parsed by this import 
-        TreeSet<String> locallyImported = new TreeSet<>(aAlreadyImportedFsIndexCollectionURLs);
-        locallyImported.removeAll(previouslyImported);
-        importUrlsCache.put(urlString, locallyImported);
-      }
-    }
-    aResults.addAll(Arrays.asList(desc.getFsIndexes()));
-  }
-
+  @Override
   protected XmlizationInfo getXmlizationInfo() {
     return XMLIZATION_INFO;
   }
